+++ xen/include/asm-arm/vgic.h
@@ -369,7 +369,7 @@ struct rdist_region;
 void vgic_v3_setup_hw(paddr_t dbase,
                       unsigned int nr_rdist_regions,
                       const struct rdist_region *regions,
-                      uint32_t rdist_stride, bool_t lpi_support);
+                      uint64_t rdist_stride, bool_t lpi_support);
 #endif
 bool_t vgic_is_domain_lpi(struct domain *d, unsigned int lpi);
